requirements.txt
streamlit
pinecone-client>=5.0.0
google-genai>=0.3.0
python-dotenv